Improved Zip Archive Build
Previously the zip build would just add the assets and binaries directly to the zip file. This means when you extract it, it's not in one folder it extracts the assets folder, binaries and what not individually. Traditionally in a zip or tarball release, all the assets are placed in a root folder with the name of the release. Look at Firefox Nightly releases for Linux for example. This fixes it by creating a root folder based on the release version then moving the assets and binaries into it, then archiving that folder.